Refactor event channel management code
Fairly complete refactor of EventChannel and EmbedUpdater (which now
ChannelUpdater and handles only a single channel). The implementation
details are hidden from EmbedManager now, instead of it needing to send
messages between EventChannel and EmbedUpdater. Now, EventChannel gets
fed event changes and sends them off to a background Tokio task for
processing. The event state for each channel is now wrapped by
ChannelEvents and this state is managed in the same task (per channel)
where the channel updates it generates are applied.

The major benefit of this refactor is that we no longer assume that the
channel's current messages match our event state. Instead,
ChannelUpdater reads the current messages on creation and figures out
what updates are needed to get them in sync with the current
ChannelEvents. Furthermore, this lets us recover on errors or when our
state is inconsistent; just recreate the ChannelUpdater, which will
fetch the channel's messages again.